feat(frontend): event subscriptions extracted — Triggers is off antd
Last of the three Triggers sections. Triggers.tsx now composes two shared
sections and one thin OSS binding, which exists only because the connections
list needs AlertPopup.

The Subscribe button's tooltip moves onto a wrapping span: a disabled button
fires no pointer events, so antd's Tooltip-on-disabled-child behaviour has to be
reproduced deliberately or the "Connect an app first" hint never appears — which
is the one case where it matters.

Also drops `fixed: "left"` from the columns carried over in all three sections.
It was an antd frozen-column setting; DataTable does not freeze, so it was inert.

Gates: lint 24/24, tsc 0 across settings-ui, oss, ee and mobile.
