+++ web/packages/agenta-shared/src/state/index.ts
@@ -5,6 +5,11 @@
 export {projectIdAtom, setProjectIdAtom} from "./project"
 export {sessionAtom, setSessionAtom} from "./session"
 export {userAtom, setUserAtom} from "./user"
+export {
+    activeUserIdAtom,
+    agentVoiceInputEnabledAtom,
+    playgroundInspectorEnabledAtom,
+} from "./featureFlags"
 export {simulatedAgentRunAtomFamily} from "./simulatedAgentRun"
 export type {SimulatedAgentRunRequest} from "./simulatedAgentRun"
 export {openAgentConfigSectionAtom} from "./openConfigSection"
